+++ linphone-app/assets/languages/fr_FR.ts
@@ -706,6 +706,16 @@ URL du serveur non configurée.</translation>
         <extracomment>&apos;Reply&apos; : Reply to a message from menu</extracomment>
         <translation>Répondre</translation>
     </message>
+    <message>
+        <source>menuViewContact</source>
+        <extracomment>&apos;View contact&apos; : Menu item to view the contact.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>menuAddContact</source>
+        <extracomment>&apos;Add to contacts&apos; : Menu item to add the contact to address book.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>ChatNoticeModel</name>