+++ linphone-app/assets/languages/zh_CN.ts
@@ -704,6 +704,16 @@
         <extracomment>&apos;Reply&apos; : Reply to a message from menu</extracomment>
         <translation>回复</translation>
     </message>
+    <message>
+        <source>menuViewContact</source>
+        <extracomment>&apos;View contact&apos; : Menu item to view the contact.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>menuAddContact</source>
+        <extracomment>&apos;Add to contacts&apos; : Menu item to add the contact to address book.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
 </context>
 <context>
     <name>ChatNoticeModel</name>