+++ linphone-app/ui/views/App/Main/Conversation.qml
@@ -665,6 +665,12 @@ ColumnLayout  {
 				updateFilter()
 			}
 		}
+		onAddContactClicked: window.setView('ContactEdit', {
+													  sipAddress: contactAddress
+												  })
+		onViewContactClicked: window.setView('ContactEdit', {
+													  sipAddress: contactAddress
+												  })
 	}
 	
 	Connections {