+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Accept/decline CLI commands.
 - Colored Emojis with its own font family.
 - OAuth2 connection to retrieve remote provisioning (Experimental and not usable without configuration).
+- Add/View contact from a message.
 
 ## 5.0.11 - undefined